move the old service impl stuff into client b/c its only for the client now
