--- rodbus/src/service/serialization/mod.rs
@@ -1,2 +1,0 @@
-mod parse;
-mod serialize;
